backend/requirements.txt
torch==1.6.0
miditoolkit
flask
flask-cors
flask-pymongo
pymongo
pydub
midi2audio
python-dotenv
shortuuid
numpy